// src/lib/grpc/jobworkerp/data/worker.ts
export interface WorkerData {
  /**
   * # Worker Name
   * Unique identifier for the worker
   */
  name: string;
  /**
   * # Description
   * Description of the worker
   */
  description: string;
  /**
   * # Runner ID
   * Reference to the runner implementation to use
   */
  runnerId:
    | RunnerId
    | undefined;
  /**
   * # Runner Settings
   * Configuration bytes for the specific runner (e.g. docker runner settings)
   */
  runnerSettings: Uint8Array;
  /**
   * # Retry Policy
   * Configuration for job retry behavior on failure
   */
  retryPolicy:
    | RetryPolicy
    | undefined;
  /**
   * # Periodic Interval
   * Milliseconds between periodic job executions (0: non-periodic)
   * If enabled, direct_response cannot be used
   */
  periodicInterval: number;
  /**
   * # Channel Name
   * Message channel for job (if empty, default channel will be used)
   */
  channel?:
    | string
    | undefined;
  /**
   * # Queue Type
   * 0: normal (no backup to database)
   * 1: with backup to database
   * 2: only to database (not in memory queue)
   */
  queueType: QueueType;
  /**
   * # Response Type
   * 0: no result
   * 1: direct response
   */
  responseType: ResponseType;
  /**
   * # Store Success Flag
   * If true, job results will be stored in database when execution succeeds
   */
  storeSuccess: boolean;
  /**
   * # Store Failure Flag
   * If true, job results will be stored in database when execution fails
   */
  storeFailure: boolean;
  /**
   * # Static Execution Flag
   * If true, fixed number instance of this worker will run at a time (with
   * lock)
   */
  useStatic: boolean;
  /**
   * # Output Broadcast Flag
   * Broadcast job results to all result listeners or not.
   * If false, cannot use JobResultService/{Listen, ListenStream,
   * ListenByWorker}.
   */
  broadcastResults: boolean;
}